# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Insider Messaging API",
        "description": "API for automated message sending system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server (port 8000)"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/messages/sent": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "List sent messages (external message IDs)",
                "description": "Returns a list of external message IDs for successfully sent messages",
                "operationId": "de24f581f8f854bb88fdf4a37763395b",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved sent message IDs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Messages retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "webhook-68c080d5bff20"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "from_database": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "from_cache": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "cache_ratio": {
                                                    "type": "number",
                                                    "example": 66.67
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-09-09T19:32:37.786252Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Messages",
            "description": "Messages"
        }
    ]
}